refactor global methods + typedef
