+++ src/server/mem_db.ts
@@ -1,0 +1,6 @@
+// TODO: Add types to these keys.
+export const shares = {};
+export const signers = {};
+export const rngs = {};
+export const addressBook = {};
+export const precomputes = {};
